port config to 1.21 for Cache Timeout

+++ src/main/java/fi/dy/masa/tweakeroo/data/ServerDataSyncer.java
@@ -209,7 +209,7 @@ public class ServerDataSyncer implements IClientTickHandler
 
     private long getCacheTimeout()
     {
-        return (long) (MathHelper.clamp(Configs.Generic.SERVER_DATA_SYNC_CACHE_TIMEOUT.getFloatValue(), 0.25f, 25.0f) * 1000L);
+        return (long) (MathHelper.clamp(Configs.Generic.SERVER_DATA_SYNC_CACHE_TIMEOUT.getDoubleValue(), 0.25f, 25.0f) * 1000L);
     }
 
     private void tickCache()
